Move GetCustomKubeConfigDirectory
