+++ src/common/app-paths/get-custom-kube-config-directory.injectable.ts
@@ -3,8 +3,8 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import directoryForKubeConfigsInjectable from "../directory-for-kube-configs.injectable";
-import joinPathsInjectable from "../../path/join-paths.injectable";
+import directoryForKubeConfigsInjectable from "./directory-for-kube-configs.injectable";
+import joinPathsInjectable from "../path/join-paths.injectable";
 
 const getCustomKubeConfigDirectoryInjectable = getInjectable({
   id: "get-custom-kube-config-directory",
